Remove Hyper-V driver code from darwin and linux targets
