+++ pkg/drivers/hyperv/powershell_windows.go
